First attempt to convert to SHACL ValidationReport...refactoring SHACL abstract syntax representation
