+++ modules/shacl/src/main/scala/es/weso/shacl/NodeKindType.scala
@@ -1,0 +1,26 @@
+package es.weso.shacl
+
+import es.weso.rdf.nodes.IRI
+import es.weso.shacl.SHACLPrefixes._
+
+sealed trait NodeKindType {
+  def id: IRI
+}
+case object IRIKind extends NodeKindType {
+  override def id = sh_IRI
+}
+case object LiteralKind extends NodeKindType {
+  override def id = sh_Literal
+}
+case object BlankNodeKind extends NodeKindType {
+  override def id = sh_BlankNode
+}
+case object BlankNodeOrIRI extends NodeKindType {
+  override def id = sh_BlankNodeOrIRI
+}
+case object BlankNodeOrLiteral extends NodeKindType {
+  override def id = sh_BlankNodeOrLiteral
+}
+case object IRIOrLiteral extends NodeKindType {
+  override def id = sh_IRIOrLiteral
+}
