+++ modules/shacl/src/main/scala/es/weso/shacl/ShapeRef.scala
@@ -1,0 +1,7 @@
+package es.weso.shacl
+
+import es.weso.rdf.nodes.RDFNode
+
+case class ShapeRef(id: RDFNode) extends AnyVal {
+  def showId = id.toString
+}
