+++ modules/shacl/src/main/scala/es/weso/shacl/validator/CheckResult.scala
@@ -3,6 +3,7 @@ package es.weso.shacl.validator
 import cats._
 import cats.implicits._
 
+
 case class CheckResult[E: Show, A: Show, Log: Show](r: (Log, Either[E, A])) {
 
   def result: Either[E, A] = r._2
