+++ modules/shacl/src/main/scala/es/weso/shacl/validator/TestValidator.scala
@@ -1,0 +1,17 @@
+package es.weso.shacl.validator
+
+import es.weso.rdf.nodes.RDFNode
+import es.weso.rdf.path.SHACLPath
+import es.weso.shacl.Shape
+import SHACLChecker._
+
+object TestValidator {
+  type CheckTyping = Check[ShapeTyping]
+  type PropertyChecker = (Attempt, SHACLPath) => CheckTyping
+  type NodeChecker = Attempt => RDFNode => CheckTyping
+  type ShapeChecker = Shape => CheckTyping
+  type NodeShapeChecker = (RDFNode, Shape) => CheckTyping
+
+  def getTyping: Check[ShapeTyping] = getEnv
+
+}
