+++ modules/shacl/src/main/scala/es/weso/shacl/validator/ViolationError.scala
@@ -1,6 +1,5 @@
 package es.weso.shacl.validator
 
-import cats.Show
 import es.weso.rdf.nodes._
 import es.weso.rdf.path.{PredicatePath, SHACLPath}
 import es.weso.shacl.SHACLPrefixes.sh
@@ -18,7 +17,6 @@ case class ViolationError(
 }
 
 object ViolationError {
-
 
  def basic(suffix: String, focusNode: RDFNode, attempt: Attempt, msg: String) =
     ViolationError(
