+++ modules/typing/src/main/scala/es/weso/typing/Typing.scala
@@ -6,6 +6,8 @@ abstract class Typing[Key, Value, Err, Evidence] {
 
   type Evidences = List[Evidence]
 
+  def allOk: Boolean = ???
+
   def hasType(key: Key, value: Value): Boolean =
     getOkValues(key) contains value
